feat: relocated <Header /> component to shared components
