+++ src/components/Header/index.tsx
@@ -12,7 +12,7 @@ import {
 } from '@mui/icons-material';
 import classnames from 'classnames';
 
-import { CustomThemeContext } from '../../../theme/CustomThemeProvider';
+import { CustomThemeContext } from '../../theme/CustomThemeProvider';
 
 import useStyles from './styles';
 import HeaderMenu from './HeaderMenu';
